packages/server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
qiskit==1.0.0
qiskit-aer==0.13.3
qiskit-algorithms==0.3.0
qiskit-optimization==0.6.1
numpy==1.24.3
python-dotenv==1.0.0
gunicorn==21.2.0
